Bug Fix
Logic from home page to grab session info from localStorage
and then fetch user data, has been moved to the app.vue
component.

The reason for this is that before, if a refresh occurred,
unless you're on the home page, then the state of the user
and it's session is lost.