Major Changes.
The handling of light and dark mode separated into
its own onMounted block.

fetchAccountById was defined inside the mounted
block. To reduce the bloat of this method, and to
make it so that logic can be reused, it was
seperated into its own component.

Additionally, if the session doesn't exist, all
the local data for it is destroyed. If the session
is expired, the same operation occurs.