Navigation guards added.
Navigation guard added to check
before entering each route that
the user is authenticated to make
sure that they are not accessing
something protected.

isAuthenticated was added to the
state management layer to check if
the session exists, and if the
session that comes back from the
server contains a valid expiration
date.