BUG FIX: User card missing data on login.
When the user logs in, an emit
method was defined to pass the
users username up to the App.vue
component.

App.vue was provided a listener
for this event and a handler was
written. This handler sets the
data locally so that there is a
value to show in the user card.